Update functionality with entered numbers etc.
Updated the algorithm so that it can accept 2+ figure values by
checking if op, a, and b are defined. Made it so that the display only
updates on the press of a second operator, equals or clear so that b
can be a 2+ figure value. Will convert the embedded code to functions
for simplicity, debugging, and readability.